MSVC warning
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@23629 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/qt4/GuiDocument.cpp
@@ -1692,7 +1692,7 @@ void GuiDocument::apply(BufferParams & params)
 	// Embedded files
 	vector<string> & files = params.extraEmbeddedFiles();
 	files.clear();
-	for (size_t i = 0; i < embeddedFilesModule->extraLW->count(); ++i) {
+	for (int i = 0; i < embeddedFilesModule->extraLW->count(); ++i) {
 		QListWidgetItem * item = embeddedFilesModule->extraLW->item(i);
 		files.push_back(fromqstr(item->text()));
 	}
